fixed wrong validation
